Break out axes and params
