+++ src/boilerdata/axes_enum.py
@@ -5,7 +5,7 @@ from enum import auto
 from boilerdata.enums import GetNameEnum
 
 
-class Axes(GetNameEnum):
+class AxesEnum(GetNameEnum):
     trial = auto()
     run = auto()
     time = auto()
