+++ src/boilerdata/pipeline.py
@@ -13,8 +13,10 @@ from propshop.library import Mat, Prop
 from scipy.constants import convert_temperature
 from scipy.stats import linregress, norm
 
-from boilerdata.axes import Axes as A  # noqa: N817
-from boilerdata.models.project import Project, Trial, get_names, get_project, set_dtypes
+from boilerdata.axes_enum import AxesEnum as A  # noqa: N814
+from boilerdata.models.axes import get_names
+from boilerdata.models.common import set_dtypes
+from boilerdata.models.project import Project, Trial, get_project
 from boilerdata.validation import validate_df, validate_runs_df
 
 # * -------------------------------------------------------------------------------- * #
